requirements.txt
Django~=5.0.3
graphene-django==3.2.0
graphene~=3.3
django-cors-headers==3.6.0
djangorestframework~=3.15.1